projects/prompt_opt/requirements.txt
pyyaml>=6.0
requests>=2.31.0
matplotlib>=3.7
